Refactor routes and static assets by step
