+++ static/js/step3.js
@@ -1,0 +1,3 @@
+$(document).ready(function(){
+    console.log('Step 3 placeholder loaded');
+});
