+++ templates/index.html
@@ -55,6 +55,8 @@
     <h2>STEP 3: (placeholder)</h2>
 </div>
 
-<script src="{{ url_for('static', filename='js/main.js') }}"></script>
+<script src="{{ url_for('static', filename='js/step1.js') }}"></script>
+<script src="{{ url_for('static', filename='js/step2.js') }}"></script>
+<script src="{{ url_for('static', filename='js/step3.js') }}"></script>
 </body>
 </html>
